Consolidated messaging nodes to a single node
