--- nodes/mqrest-get.html
@@ -1,127 +1,0 @@
-<!--
-  Copyright 2022 IBM Corp.
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-
-<script type="text/html" data-template-name="mqrest-get">
-  <div class="form-row">
-    <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
-    <input type="text" id="node-input-name" placeholder="Name">
-  </div>
-
-  <div class="form-row">
-    <label for="node-input-user"><i class="fa fa-folder-close"></i> User</label>
-    <select id="node-input-user" style="width:68%;">
-    </select>
-  </div>
-
-  <div class="form-row">
-    <label for="node-input-server"><i class="fa fa-folder-close"></i> Server</label>
-    <select id="node-input-server" style="width:68%;">
-    </select>
-  </div>
-
-  <div class="form-row">
-    <label for="node-input-apiv"><i class="fa fa-folder-close"></i> API Version</label>
-    <input type="text" id="node-input-apiv">
-    </select>
-  </div>
-
-  <div class="form-row">
-    <label for="node-input-op"><i class="fa fa-folder-close"></i> Browse/Delete</label>
-    <input type="text" id="node-input-op">
-  </div>
-
-  <div class="form-tips">
-    <span>This node supports the following parameters via an inject node:</span>
-    <br>
-    <b>Required:</b>
-    <ul>
-      <li><b>msg.qmgr</b> : The name of the MQ Queue Manager.</li>
-      <li><b>msg.qname</b> : The name of the Queue.</li>
-    </ul>
-  </div>
-
-</script>
-
-<script type="text/html" data-help-name="mqrest-get">
-<p>This node makes use of the MQ REST API to read with browse or delete messages from a MQ Queue. Use the
-configuration to specify the full API URL.</p>
-
-<p>The message if found is placed on <code>msg.payload</code>. If the message is a JSON object, then
-it is first converted into a JSON object before placing on <code>msg.payload</code>.</p>
-
-</script>
-
-<style type="text/css">
-  .node_label_white {
-    fill: white;
-  }
-
-  #palette_node_mqrest>div.palette_label {
-    color: white;
-  }
-</style>
-
-<script type="text/javascript">
-  (function () {
-    RED.nodes.registerType('mqrest-get', {
-      category: 'IBM MQ REST',
-      defaults: {
-        name: { value: '' },
-        user: { value: "", type: "user-config" },
-        server: { value: "", type: "server-config" },
-        apiv: { value: "v2" },
-        op: { value: "", required: true }
-      },
-      credentials: {
-      },
-      color: '#C7E9C0',
-      inputs: 1,
-      outputs: 1,
-      paletteLabel: 'MQ GET',
-      icon: 'feed.png',
-      label: function () {
-        return this.name || 'MQ GET';
-      },
-      labelStyle: function () {
-        return this.name ? 'node_label_italic node_label_black' : 'node_label_black';
-      },
-      oneditprepare: function () {
-        $("#node-input-op").typedInput({
-          types: [
-            {
-              value: "Browse/Delete",
-              options: [
-                { value: "GET", label: "Browse" },
-                { value: "DELETE", label: "Delete" },
-              ]
-            }
-          ]
-        }),
-        $("#node-input-apiv").typedInput({
-          types: [
-            {
-              value: "version",
-              options: [
-                { value: "v1", label: "1"},
-                { value: "v2", label: "2"}
-              ]
-            }
-          ]
-        })
-      }
-    });
-  })();
-</script>
