--- nodes/mqrest-put.html
@@ -1,125 +1,0 @@
-<!--
-  Copyright 2022 IBM Corp.
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-
-<script type="text/html" data-template-name="mqrest-put">
-  <div class="form-row">
-    <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
-    <input type="text" id="node-input-name" placeholder="Name">
-  </div>
-
-  <div class="form-row">
-    <label for="node-input-user"><i class="fa fa-folder-close"></i> User</label>
-    <select id="node-input-user" style="width:68%;">
-    </select>
-  </div>
-
-  <div class="form-row">
-    <label for="node-input-server"><i class="fa fa-folder-close"></i> Server</label>
-    <select id="node-input-server" style="width:68%;">
-    </select>
-  </div>
-
-  <div class="form-row">
-    <label for="node-input-apiv"><i class="fa fa-folder-close"></i> API Version</label>
-    <input type="text" id="node-input-apiv">
-    </select>
-  </div>
-
-  <div class="form-row">
-    <label for="node-input-contentType"><i class="fa fa-book"></i> Content Type: </label>
-    <select type="text" id="node-input-contentType" style="display: inline-block; width: 70%;">
-      <option value="text/plain">text/plain</option>
-      <option value="application/json">application/json</option>
-    </select>
-  </div>
-
-  <div class="form-tips">
-    <span>This node supports the following parameters via an inject node:</span>
-    <br>
-    <b>Required:</b>
-    <ul>
-      <li><b>msg.qmgr</b> : The name of the MQ Queue Manager.</li>
-      <li><b>msg.qname</b> : The name of the Queue.</li>
-      <li><b>msg.payload</b> : The message to put onto the queue.</li>
-    </ul>
-    Optional:
-    <ul>
-      <li><b>msg.csrf</b> : The The CSRF protection header. </li>
-    </ul>
-  </div>
-  
-
-</script>
-
-<script type="text/html" data-help-name="mqrest-put">
-  <p>This node makes use of the MQ REST API to write messages to a MQ Queue. Use the
-  configuration to specify parameters.</p>
-
-  <p>The message to post onto the queue is provided as input in <code>msg.payload</code>.</p>
-  <p>Specify the expected content type in the configuration. The node will check that
-  the contents of <code>msg.payload</code> matches the specified type.</p>
-</script>
-
-<style type="text/css">
-  .node_label_white {
-    fill: white;
-  }
-
-  #palette_node_mqrest-out>div.palette_label {
-    color: white;
-  }
-</style>
-
-<script type="text/javascript">
-  (function() {
-    RED.nodes.registerType('mqrest-put', {
-      category: 'IBM MQ REST',
-      defaults: {
-        name: { value: '' },
-        user: { value: "", type:"user-config", required:true},
-        server: { value:"", type:"server-config", required:true},
-        apiv: {value: 'v2' },
-        contentType: { value: 'text/plain' }
-      },
-      credentials: {
-      },
-      color: '#C7E9C0',
-      inputs: 1,
-      outputs: 0,
-      paletteLabel: 'MQ PUT',
-      icon: 'feed.png',
-      label: function() {
-        return this.name || 'MQ PUT';
-      },
-      labelStyle: function() {
-        return this.name ? 'node_label_italic node_label_black' : 'node_label_black';
-      },
-      oneditprepare: function () {
-          $("#node-input-apiv").typedInput({
-            types: [
-              {
-                value: "version",
-                options: [
-                  { value: "v1", label: "1"},
-                  { value: "v2", label: "2"}
-                ]
-              }
-            ]
-          })
-        }
-    });
-  })();
-</script>
